requirements.txt
asgiref==3.8.1
colorama==0.4.6
Django==4.2.18
django-formtools==2.5.1
django-otp==1.5.4
django-phonenumber-field==8.0.0
django-two-factor-auth==1.17.0
djangorestframework==3.15.2
phonenumbers==8.13.54
pypng==0.20220715.0
qrcode==7.4.2
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
